Temp hax to point GeoData at codfw DC
The eqiad elasticsearch cluster is currently running a mixed version
cluster, having both versions 5.1.2 and 5.3.2. While elasticsearch is
supposed to guarnatee that mixed clusters of the same major version will
work fine, it seems they made a mistake and there is an enum related to
geo search that was changed. This basically means requests from one
version to the other fail because the enum's get interpreted wrong.

This will fix itself as the cluster upgrade is completed. Pushing only
to wmf.4 (and already in wmf.2) so prevent errors as this sees 5 or 10k
uses per hour.

Change-Id: Ibe32c3f1a15e03583edbd42c1b53704b467ffe9b